+++ spec/features/staff/users_can_view_reports_spec.rb
@@ -140,7 +140,7 @@ RSpec.feature "Users can view reports" do
       end
     end
 
-    scenario "can download a CSV of the report" do
+    scenario "they see helpful guidance about and can download a CSV of their own report" do
       report = create(:report, :active)
 
       visit reports_path
@@ -149,36 +149,14 @@ RSpec.feature "Users can view reports" do
         click_on t("default.link.show")
       end
 
-      click_on t("action.report.download.button")
+      expect(page).to have_content("Download a CSV file to review offline.")
+      expect(page).to have_link("guidance in the help centre (opens in new tab)")
+
+      click_link t("action.report.download.button")
 
       expect(page.response_headers["Content-Type"]).to include("text/csv")
       header = page.response_headers["Content-Disposition"]
       expect(header).to match(/#{ERB::Util.url_encode("#{report.organisation.beis_organisation_reference}-report.csv")}\z/)
-    end
-
-    scenario "can download a spending breakdown CSV of the report" do
-      fund = create(:fund_activity)
-      programme = create(:programme_activity, parent: fund)
-      dp_org = create(:delivery_partner_organisation)
-      project = create(:project_activity, organisation: dp_org, parent: programme)
-      report = create(:report, :active, organisation: dp_org, fund: programme.parent)
-      create(:actual, report: report, parent_activity: project)
-
-      visit reports_path
-
-      within "##{report.id}" do
-        click_on t("default.link.show")
-      end
-
-      click_on t("action.report.spending_download.button")
-
-      expect(page.response_headers["Content-Type"]).to include("text/csv")
-      header = page.response_headers["Content-Disposition"]
-      expect(header).to match(/#{ERB::Util.url_encode("#{report.organisation.beis_organisation_reference}-report.csv")}\z/)
-
-      download = CSV.parse(page.body)
-      expected_row = ActivitySpendingBreakdown.new(report: report, activity: project).values
-      expect(download.to_a).to include(expected_row)
     end
 
     context "if the report description is empty" do
@@ -229,10 +207,6 @@ RSpec.feature "Users can view reports" do
 
       it "they can be downloaded as CSV" do
         click_on "Download report as CSV file"
-      end
-
-      it "the spending breakdown can be downloaded" do
-        click_on "Download spending breakdown as CSV file"
       end
     end
   end
@@ -389,34 +363,18 @@ RSpec.feature "Users can view reports" do
         expect(page).to have_http_status(401)
       end
 
-      scenario "can download a CSV of their own report" do
+      scenario "they see helpful guidance about and can download a CSV of their own report" do
         report = create(:report, :active, organisation: delivery_partner_user.organisation)
 
         visit reports_path
-
         within "##{report.id}" do
           click_on t("default.link.show")
         end
 
-        click_on t("action.report.download.button")
+        expect(page).to have_content("Download a CSV file to review offline.")
+        expect(page).to have_link("guidance in the help centre (opens in new tab)")
 
-        expect(page.response_headers["Content-Type"]).to include("text/csv")
-
-        expect(page.response_headers["Content-Type"]).to include("text/csv")
-        header = page.response_headers["Content-Disposition"]
-        expect(header).to match(ERB::Util.url_encode("#{report.organisation.beis_organisation_reference}-report.csv"))
-      end
-
-      scenario "can download a spending breakdown CSV of their own report" do
-        report = create(:report, :active, organisation: delivery_partner_user.organisation)
-
-        visit reports_path
-
-        within "##{report.id}" do
-          click_on t("default.link.show")
-        end
-
-        click_on t("action.report.spending_download.button")
+        click_link t("action.report.download.button")
 
         expect(page.response_headers["Content-Type"]).to include("text/csv")